feat(npm)!: consolidate ui and composable under the vizejs scope
Renames the four first-publish packages out of the unprovisioned
vizeui/vizecompose scopes and folds the single-module packages into
their group cores, so the public surface is two packages instead of
four: @vizejs/ui (with media, media/pdf, and media/source as lazy
entries alongside the existing component entries) and
@vizejs/composable (with temporal as a lazy entry carrying the
polyfill dependency, keeping the root entry dependency-free). The
private tooling package follows as @vizejs/ui-tooling.

Distribution tests move with their modules and now pin the merged
subpath exports, per-entry gzip budgets cover the new entries, the
release workflow publishes exactly the two packages, compose/core
joins the checked, packed, and tested CI package lists it was missing
from, and the preflight inventory expectation drops the two retired
manifests. Nested package dist output joins the gitignore so builds
can never block the release preflight again, and the native-binaries
lockfile catalog keeps the 0.299.1 pins that a plain install drops
while those platform packages are still publishing.
