+++ .gitignore
@@ -8,6 +8,7 @@ _build/
 
 # npm package build outputs
 npm/*/dist/
+npm/*/*/dist/
 editors/*/dist/
 npm/builder-vite/wasm/
 editors/vscode/.vscode-test/
