+++ npm/compose/core/src/temporal.test.ts
@@ -8,7 +8,7 @@ import {
   useTemporalNow,
   useTemporalZonedDateTime,
   type TemporalScheduler,
-} from "./index.ts";
+} from "./temporal.ts";
 
 interface ScheduledTimer {
   readonly callback: () => void;
@@ -142,14 +142,14 @@ void test("publishes an importable ESM distribution with declarations", async ()
     await readFile(new URL("../package.json", import.meta.url), "utf8"),
   ) as {
     readonly exports: {
-      readonly ".": { readonly import: string; readonly types: string };
+      readonly "./temporal": { readonly import: string; readonly types: string };
     };
   };
 
-  assert.equal(packageJson.exports["."].import, "./dist/index.mjs");
-  assert.equal(packageJson.exports["."].types, "./dist/index.d.mts");
+  assert.equal(packageJson.exports["./temporal"].import, "./dist/temporal.mjs");
+  assert.equal(packageJson.exports["./temporal"].types, "./dist/temporal.d.mts");
 
-  const distribution = await import("../dist/index.mjs");
+  const distribution = await import("../dist/temporal.mjs");
   assert.equal(typeof distribution.useTemporalNow, "function");
   assert.equal(typeof distribution.Temporal.Instant.from, "function");
 });
