--- npm/compose/temporal/vite.config.ts
@@ -1,18 +1,0 @@
-import { defineConfig } from "vite-plus";
-
-export default defineConfig({
-  lint: {
-    ignorePatterns: ["dist/**"],
-    options: { typeAware: true },
-  },
-  fmt: { ignorePatterns: ["dist/**"] },
-  pack: {
-    entry: ["src/index.ts"],
-    format: "esm",
-    dts: true,
-    clean: true,
-    deps: {
-      neverBundle: ["vue", "temporal-polyfill-lite"],
-    },
-  },
-});
