+++ npm/ui/core/vite.config.ts
@@ -16,6 +16,9 @@ export default defineConfig({
       "controllable-state": "src/controllable-state.ts",
       primitive: "src/primitive.ts",
       "visually-hidden": "src/visually-hidden.ts",
+      media: "src/media.ts",
+      "media-pdf": "src/pdf.ts",
+      "media-source": "src/media-source.ts",
     },
     format: "esm",
     dts: { vue: true },
