--- npm/ui/media/package.json
@@ -1,67 +1,0 @@
-{
-  "name": "@vizeui/media",
-  "version": "0.299.1",
-  "description": "Accessible, type-safe media foundations for Vize applications",
-  "keywords": [
-    "accessibility",
-    "audio",
-    "media",
-    "pdf",
-    "video"
-  ],
-  "homepage": "https://github.com/ubugeeei-prod/vize",
-  "bugs": {
-    "url": "https://github.com/ubugeeei-prod/vize/issues"
-  },
-  "license": "MIT",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/ubugeeei-prod/vize.git",
-    "directory": "npm/ui/media"
-  },
-  "files": [
-    "dist"
-  ],
-  "type": "module",
-  "sideEffects": false,
-  "main": "./dist/index.mjs",
-  "types": "./dist/index.d.mts",
-  "exports": {
-    ".": {
-      "types": "./dist/index.d.mts",
-      "import": "./dist/index.mjs",
-      "default": "./dist/index.mjs"
-    },
-    "./pdf": {
-      "types": "./dist/pdf.d.mts",
-      "import": "./dist/pdf.mjs",
-      "default": "./dist/pdf.mjs"
-    },
-    "./source": {
-      "types": "./dist/source.d.mts",
-      "import": "./dist/source.mjs",
-      "default": "./dist/source.mjs"
-    }
-  },
-  "publishConfig": {
-    "access": "public"
-  },
-  "scripts": {
-    "build": "vp pack",
-    "dev": "vp pack --watch",
-    "pretest": "vp pack && pnpm check:size",
-    "test": "vp exec node --test 'src/**/*.test.ts'",
-    "check": "vp check src scripts vite.config.ts",
-    "check:fix": "vp check --fix src scripts vite.config.ts",
-    "check:size": "node scripts/check-size.mjs",
-    "fmt": "vp fmt --write src scripts vite.config.ts"
-  },
-  "devDependencies": {
-    "@types/node": "catalog:typescript",
-    "typescript": "catalog:typescript",
-    "vite-plus": "catalog:vite-stack"
-  },
-  "engines": {
-    "node": ">=24"
-  }
-}
