+++ tests/tooling/release-preflight-runner.test.ts
@@ -100,7 +100,6 @@ test("release metadata inventory discovers every non-private npm and editor pack
       "npm/builder/vite/package.json",
       "npm/cli/package.json",
       "npm/compose/core/package.json",
-      "npm/compose/temporal/package.json",
       "npm/framework/musea-nuxt/package.json",
       "npm/framework/nuxt/package.json",
       "npm/fresco-native/package.json",
@@ -110,7 +109,6 @@ test("release metadata inventory discovers every non-private npm and editor pack
       "npm/native/package.json",
       "npm/oxint/package.json",
       "npm/ui/core/package.json",
-      "npm/ui/media/package.json",
       "npm/wasm/package.json",
     ],
   );
